Renamed the User test file to be consistend and moved setup and teardown logic into a base test class
